+++ include/app/appinfo.h
@@ -39,190 +39,194 @@ namespace Nickvision::App
         /**
          * @brief Constructs an AppInfo.
          */
-        AppInfo() = default;
+        AppInfo() noexcept = default;
+        AppInfo(const AppInfo&) noexcept = default;
+        AppInfo(AppInfo&&) noexcept = default;
         /**
          * @brief Constructs an AppInfo.
          * @param id The application id
          * @param name The application name
          * @param englishShortName The application short name in English
          */
-        AppInfo(const std::string& id, const std::string& name, const std::string& englishShortName);
+        AppInfo(const std::string& id, const std::string& name, const std::string& englishShortName) noexcept;
         /**
          * @brief Gets the application id.
          * @return The application id
          */
-        const std::string& getId() const;
+        const std::string& getId() const noexcept;
         /**
          * @brief Sets the application id.
          * @param id The application id
          */
-        void setId(const std::string& id);
+        void setId(const std::string& id) noexcept;
         /**
          * @brief Gets the application name.
          * @return The application name
          */
-        const std::string& getName() const;
+        const std::string& getName() const noexcept;
         /**
          * @brief Sets the application name.
          * @param name The application name
          */
-        void setName(const std::string& name);
+        void setName(const std::string& name) noexcept;
         /**
          * @brief Gets the application short name.
          * @return The application short name
          */
-        const std::string& getShortName() const;
+        const std::string& getShortName() const noexcept;
         /**
          * @brief Sets the application short name.
          * @param shortName The application short name
          */
-        void setShortName(const std::string& shortName);
+        void setShortName(const std::string& shortName) noexcept;
         /**
          * @brief Gets the application English short name.
          * @return The application English short name
          */
-        const std::string& getEnglishShortName() const;
+        const std::string& getEnglishShortName() const noexcept;
         /**
          * @brief Sets the application English short name.
          * @param englishShortName The application English short name
          */
-        void setEnglishShortName(const std::string& englishShortName);
+        void setEnglishShortName(const std::string& englishShortName) noexcept;
         /**
          * @brief Gets the application description.
          * @return The application description
          */
-        const std::string& getDescription() const;
+        const std::string& getDescription() const noexcept;
         /**
          * @brief Sets the application description.
          * @param description The application description
          */
-        void setDescription(const std::string& description);
+        void setDescription(const std::string& description) noexcept;
         /**
          * @brief Gets the application version.
          * @return The application version
          */
-        const Update::Version& getVersion() const;
+        const Update::Version& getVersion() const noexcept;
         /**
          * @brief Sets the application version.
          * @param version The application version
          */
-        void setVersion(const Update::Version& version);
+        void setVersion(const Update::Version& version) noexcept;
         /**
          * @brief Gets the application changelog.
          * @return The application changelog
          */
-        const std::string& getChangelog() const;
+        const std::string& getChangelog() const noexcept;
         /**
          * @brief Sets the application changelog.
          * @param changelog The application changelog
          */
-        void setChangelog(const std::string& changelog);
+        void setChangelog(const std::string& changelog) noexcept;
         /**
          * @brief Gets the application changelog in HTML form
          * @return The application changelog in HTML
          */
-        const std::string& getHtmlChangelog() const;
+        const std::string& getHtmlChangelog() const noexcept;
         /**
          * @brief Gets the application source repo url
          * @return The application source repo url
          */
-        const std::string& getSourceRepo() const;
+        const std::string& getSourceRepo() const noexcept;
         /**
          * @brief Sets the application source repo url
          * @param sourceRepo The application source repo url
          * @return True if the new source repo was set, else false
          * @return A false return value means that the sourceRepo param was not a valid formatted URL
          */
-        bool setSourceRepo(const std::string& sourceRepo);
+        bool setSourceRepo(const std::string& sourceRepo) noexcept;
         /**
          * @brief Gets the application issue tracker url
          * @return The application issue tracker url
          */
-        const std::string& getIssueTracker() const;
+        const std::string& getIssueTracker() const noexcept;
         /**
          * @brief Sets the application issue tracker url
          * @param issueTracker The application issue tracker url
          * @return True if the new issue tracker was set, else false
          * @return A false return value means that the issueTracker param was not a valid formatted URL
          */
-        bool setIssueTracker(const std::string& issueTracker);
+        bool setIssueTracker(const std::string& issueTracker) noexcept;
         /**
          * @brief Gets the application support url
          * @return The application support url
          */
-        const std::string& getSupportUrl() const;
+        const std::string& getSupportUrl() const noexcept;
         /**
          * @brief Sets the application support url
          * @param supportUrl The application support url
          * @return True if the new support url was set, else false
          * @return A false return value means that the supportUrl param was not a valid formatted URL
          */
-        bool setSupportUrl(const std::string& supportUrl);
+        bool setSupportUrl(const std::string& supportUrl) noexcept;
         /**
          * @brief Gets the application html docs store url
          * @return The application html docs store url
          */
-        const std::string& getHtmlDocsStore() const;
+        const std::string& getHtmlDocsStore() const noexcept;
         /**
          * @brief Sets the application html docs store url
          * @param htmlDocsStore The application html docs store url
          */
-        void setHtmlDocsStore(const std::string& htmlDocsStore);
+        void setHtmlDocsStore(const std::string& htmlDocsStore) noexcept;
         /**
          * @brief Gets the extra links for the application
          * @return The extra links for the application
          */
-        std::unordered_map<std::string, std::string>& getExtraLinks();
+        std::unordered_map<std::string, std::string>& getExtraLinks() noexcept;
         /**
          * @brief Gets the extra links for the application
          * @return The extra links for the application
          */
-        const std::unordered_map<std::string, std::string>& getExtraLinks() const;
+        const std::unordered_map<std::string, std::string>& getExtraLinks() const noexcept;
         /**
          * @brief Gets the application developers
          * @return The application developers
          */
-        std::unordered_map<std::string, std::string>& getDevelopers();
+        std::unordered_map<std::string, std::string>& getDevelopers() noexcept;
         /**
          * @brief Gets the application developers
          * @return The application developers
          */
-        const std::unordered_map<std::string, std::string>& getDevelopers() const;
+        const std::unordered_map<std::string, std::string>& getDevelopers() const noexcept;
         /**
          * @brief Gets the application designers
          * @return The application designers
          */
-        std::unordered_map<std::string, std::string>& getDesigners();
+        std::unordered_map<std::string, std::string>& getDesigners() noexcept;
         /**
          * @brief Gets the application designers
          * @return The application designers
          */
-        const std::unordered_map<std::string, std::string>& getDesigners() const;
+        const std::unordered_map<std::string, std::string>& getDesigners() const noexcept;
         /**
          * @brief Gets the application artists
          * @return The application artists
          */
-        std::unordered_map<std::string, std::string>& getArtists();
+        std::unordered_map<std::string, std::string>& getArtists() noexcept;
         /**
          * @brief Gets the application artists
          * @return The application artists
          */
-        const std::unordered_map<std::string, std::string>& getArtists() const;
+        const std::unordered_map<std::string, std::string>& getArtists() const noexcept;
         /**
          * @brief Gets the application translator credits
          * @return The application translator credits
          */
-        const std::string& getTranslatorCredits() const;
+        const std::string& getTranslatorCredits() const noexcept;
         /**
          * @brief Sets the application translator credits
          * @param translatorCredits The application translator credits
          */
-        void setTranslatorCredits(const std::string& translatorCredits);
+        void setTranslatorCredits(const std::string& translatorCredits) noexcept;
         /**
          * @brief Gets the list of translator names
          * @return The application translator credits
          */
-        std::vector<std::string> getTranslatorNames() const;
+        std::vector<std::string> getTranslatorNames() const noexcept;
+        AppInfo& operator=(const AppInfo&) noexcept = default;
+        AppInfo& operator=(AppInfo&&) noexcept = default;
 
     private:
         std::string m_id;
@@ -242,14 +246,6 @@ namespace Nickvision::App
         std::unordered_map<std::string, std::string> m_designers;
         std::unordered_map<std::string, std::string> m_artists;
         std::string m_translatorCredits;
-
-    public:
-        /**
-         * @brief Converts a map of URLs to a vector
-         * @param urls The map of URLs
-         * @return The vector of URLs
-         */
-        static std::vector<std::string> convertUrlMapToVector(const std::unordered_map<std::string, std::string>& urls);
     };
 }