+++ include/filesystem/filesystemchangedeventargs.h
@@ -40,17 +40,21 @@ namespace Nickvision::Filesystem
          * @param path The path of the file/folder that changed
          * @param why The action that caused the file system object to change
          */
-        FileSystemChangedEventArgs(const std::filesystem::path& path, FileAction why);
+        FileSystemChangedEventArgs(const std::filesystem::path& path, FileAction why) noexcept;
+        FileSystemChangedEventArgs(const FileSystemChangedEventArgs&) noexcept = default;
+        FileSystemChangedEventArgs(FileSystemChangedEventArgs&&) noexcept = default;
         /**
          * @brief Gets the path of the changed file system object system object.
          * @return The path of the changed file/folder
          */
-        const std::filesystem::path& getPath() const;
+        const std::filesystem::path& getPath() const noexcept;
         /**
          * @brief Gets the action that caused the file system object to change.
          * @return The action that caused the file system object to change
          */
-        FileAction getWhy() const;
+        FileAction getWhy() const noexcept;
+        FileSystemChangedEventArgs& operator=(const FileSystemChangedEventArgs&) noexcept = default;
+        FileSystemChangedEventArgs& operator=(FileSystemChangedEventArgs&&) noexcept = default;
 
     private:
         std::filesystem::path m_path;