+++ src/notifications/appnotification.cpp
@@ -6,12 +6,12 @@ namespace Nickvision::Notifications
 {
     static Event<NotificationSentEventArgs> s_sent;
 
-    Event<NotificationSentEventArgs>& AppNotification::sent()
+    Event<NotificationSentEventArgs>& AppNotification::sent() noexcept
     {
         return s_sent;
     }
 
-    void AppNotification::send(const NotificationSentEventArgs& args)
+    void AppNotification::send(const NotificationSentEventArgs& args) noexcept
     {
         s_sent.invoke(args);
     }